move code to shared.ts, fix ReflectType / TypeOf for relativeNames, fix bugs wih parsing shared.ts code
